Split EditorSplitter tests into a separate file.
